Code snippet that causes models to crash with Intel -O3
Here's  a simple case that shows how the FORTRAN optimization switch -O3 could
affect the model runs without an obvious cause. 
 
Adding the following seemingly inconsequential code snippet in the 
right place inside mom6 can result the model to produce NaNs when compiled in prod mode -O3  (which runs fine, as it should, with -O2 and -O0).   

!--------------------------------------------------------
  logical :: totallyFalse
  totallyFalse=.false.
  if (totallyFalse) then
    do k=1,nz ; do j=G%jsc,G%jec ; do i=G%isc,G%iec
       u(i,j,k) = u(i,j,k) + 0.0
    enddo ; enddo ; enddo
  endif
  if (totallyFalse) then
    do k=1,nz ; do j=G%jsc,G%jec ; do i=G%isc,G%iec
       v(i,j,k) = v(i,j,k) + 0.0
    enddo ; enddo ; enddo
  endif

!--------------------------------------------------------

This was tested with Intel16 on c3/c4 and was a root cause of a recent issue with the regression testing which Ray and I worked on to resolve.

Fortunately in this case it resulted a detectable change, i.e., NaN's followed by a crash.
Could something like this go undetected by producing round-off results? 
Is this a compiler bug?

The major lesson I learned was not to ignore  "if(false)" blocks in reading the code!